backend/requirements.txt
fastapi==0.104.1
pydantic-settings==2.1.0
uvicorn[standard]==0.24.0
singlestoredb==1.4.0
cryptography==41.0.7
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
openai==1.3.7
pandas==2.1.3
numpy==1.25.2
scikit-learn==1.3.2
redis==5.0.1
celery==5.3.4
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1
jinja2==3.1.2
python-socketio==5.10.0
websockets==12.0
schedule==1.2.0
asyncio-mqtt==0.16.1
